requirements.txt
mysql-connector==2.1.6
requests==2.20.0
urllib3==1.24